refactor(cli): split core e2e tests
